[Iter13] feat:Add retries

+++ init.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS metrics
 (
     id text NOT NULL,
     type text NOT NULL,
-    delta integer,
+    delta bigint,
     value double precision,
     CONSTRAINT metrics_pkey PRIMARY KEY (id)
 )
